diskimg: added clippy wrapper
To make invoking the different executables a bit easier I have added a
small wrapper that uses clippy to do some argument parsing. Once
arguments have been parsed (and sensible defaults assigned), it then
just calls into the main of the original executable.

The main function has C++ ABI, so has to be re-declared to be callable.
Also has to be renamed to avoid duplicate symbols.